Reorder Edge and Node to minimize struct size
Change Edge fields to be int32 when possible to reduce their size on
64bits platforms.

Reduce memory usage by 0.62% so not a lot.

name                  old time/op    new time/op    delta
CLParser                 293µs ± 0%     293µs ± 1%    ~     (p=0.684
n=10+10)
LoadManifest             483ms ± 4%     485ms ± 3%    ~     (p=0.497
n=10+9)
CanonicalizePathBits    94.2ns ± 2%    93.4ns ± 0%  -0.79%  (p=0.000
n=10+9)
CanonicalizePath        91.6ns ± 0%    91.8ns ± 0%  +0.20%  (p=0.019
n=8+10)

name                  old alloc/op   new alloc/op   delta
CLParser                 167kB ± 0%     167kB ± 0%    ~     (all equal)
LoadManifest             440MB ± 0%     437MB ± 0%  -0.64%  (p=0.000
n=9+10)
CanonicalizePathBits     80.0B ± 0%     80.0B ± 0%    ~     (all equal)
CanonicalizePath         80.0B ± 0%     80.0B ± 0%    ~     (all equal)

name                  old allocs/op  new allocs/op  delta
CLParser                 1.75k ± 0%     1.75k ± 0%    ~     (all equal)
LoadManifest             3.38M ± 0%     3.38M ± 0%    ~     (p=0.458
n=9+10)
CanonicalizePathBits      1.00 ± 0%      1.00 ± 0%    ~     (all equal)
CanonicalizePath          1.00 ± 0%      1.00 ± 0%    ~     (all equal)

+++ dyndep.go
@@ -120,7 +120,7 @@ func (d *DyndepLoader) UpdateEdge(edge *Edge, dyndeps *Dyndeps, err *string) boo
 
 	// Add the dyndep-discovered outputs to the edge.
 	edge.outputs_ = append(edge.outputs_, dyndeps.implicit_outputs_...)
-	edge.implicit_outs_ += len(dyndeps.implicit_outputs_)
+	edge.implicit_outs_ += int32(len(dyndeps.implicit_outputs_))
 
 	// Add this edge as incoming to each new output.
 	for _, i := range dyndeps.implicit_outputs_ {
@@ -139,12 +139,12 @@ func (d *DyndepLoader) UpdateEdge(edge *Edge, dyndeps *Dyndeps, err *string) boo
 
 	// Add the dyndep-discovered inputs to the edge.
 	old := edge.inputs_
-	offset := len(edge.inputs_) - edge.order_only_deps_
+	offset := len(edge.inputs_) - int(edge.order_only_deps_)
 	edge.inputs_ = make([]*Node, len(edge.inputs_)+len(dyndeps.implicit_inputs_))
 	copy(edge.inputs_, old[:offset])
 	copy(edge.inputs_[offset:], dyndeps.implicit_inputs_)
 	copy(edge.inputs_[offset+len(dyndeps.implicit_inputs_):], old[offset:])
-	edge.implicit_deps_ += len(dyndeps.implicit_inputs_)
+	edge.implicit_deps_ += int32(len(dyndeps.implicit_inputs_))
 
 	// Add this edge as outgoing from each new input.
 	for _, i := range dyndeps.implicit_inputs_ {
